\name{tscount-package}
\alias{tscount-package}
\alias{tscount}
\docType{package}
\title{
Analysis of Count Time Series
}
\description{
Collection of R functions for analysis of count time series. Currently the focus is on count time series following generalised linear models.
}
\details{
See the main function \code{\link{tsglm}} for more details on the usage of the package. There is a vignette available which introduces the functionality of the package and its underlying statistical methods (\code{vignette("tsglm", package="tscount")}).
}
\author{
Tobias Liboschik <liboschik@statistik.tu-dortmund.de>
}
\references{
Christou, V. and Fokianos, K. (2014) Quasi-likelihood inference for negative binomial time series models. \emph{Journal of Time Series Analysis} \bold{35(1)}, 55--78, \url{http://dx.doi.org/10.1002/jtsa.12050}.

Christou, V. and Fokianos, K. (2015) Estimation and testing linearity for non-linear mixed poisson autoregressions. \emph{Electronic Journal of Statistics} \bold{9}, 1357--1377, \url{http://dx.doi.org/10.1214/15-EJS1044}.

Ferland, R., Latour, A. and Oraichi, D. (2006) Integer-valued GARCH process. \emph{Journal of Time Series Analysis} \bold{27(6)}, 923--942, \url{http://dx.doi.org/10.1111/j.1467-9892.2006.00496.x}.

Fokianos, K. and Fried, R. (2010) Interventions in INGARCH processes. \emph{Journal of Time Series Analysis} \bold{31(3)}, 210--225, \url{http://dx.doi.org/10.1111/j.1467-9892.2010.00657.x}.

Fokianos, K., and Fried, R. (2012) Interventions in log-linear Poisson autoregression. \emph{Statistical Modelling} \bold{12(4)}, 299--322. \url{http://dx.doi.org/10.1177/1471082X1201200401}.

Fokianos, K., Rahbek, A. and Tjostheim, D. (2009) Poisson autoregression. \emph{Journal of the American Statistical Association} \bold{104(488)}, 1430--1439, \url{http://dx.doi.org/10.1198/jasa.2009.tm08270}.

Fokianos, K. and Tjostheim, D. (2011) Log-linear Poisson autoregression. \emph{Journal of Multivariate Analysis} \bold{102(3)}, 563--578, \url{http://dx.doi.org/10.1016/j.jmva.2010.11.002}.

Liboschik, T., Kerschke, P., Fokianos, K. and Fried, R. (2014) Modelling interventions in INGARCH processes. \emph{International Journal of Computer Mathematics} (published online), \url{http://dx.doi.org/10.1080/00207160.2014.949250}.
}

\keyword{Intervention detection}
\keyword{Model assessment}
\keyword{Inference}
\keyword{Prediction}
\keyword{Simulation}
